requirements.txt
autopep8>=1.4.3
GitPython>=2.1.11
matplotlib>=3.0.2
numpy>=1.15.4
pycparser>=2.19
pyserial>=3.4
requests>=2.21.0
urllib3>=1.24.1
jinja2>=2.10
xmltodict>=0.12.0
pylink-square>=0.2.0
